+++ src/main/java/com/example/pfairplayservice/model/enumfield/DisClosureOption.java
@@ -1,4 +1,4 @@
-package com.example.pfairplayservice.model.origin;
+package com.example.pfairplayservice.model.enumfield;
 
 public enum DisClosureOption {
